chore: Move post publisher settings endpoint to endpoints section of app
